Refactor and add some more tests
The previous implementation was relying on a bug in `puppet-lint` not
being fixed. This replacement should continue to work after it's fixed.

(We could choose to simplify the logic in this check if we depended on
the fixed version; but there would be no immediate pressure to do so.)

The `exec_attributes` helper should be a bit more robust too.  We only
consider tokens that appear before the `=>` token as attributes. This
prevents the possibility of mistaking an `unless` inside an expression
assigned to an attribute as the `unless` attribute itself. A new test
covers this.

`next_token_of` will also skip over `{ } blocks` by default. (I don't
think the `Exec` type actually has any attribute that accepts a hash
value, but this would prevent confusing those hash keys as resource
attributes.)

I've included one `pending` test that shows a current limitation. The
logic to cover this is probably too complex to be worth implementing,
and there are still other cases that we will never be able to cover with
the way puppet-lint manifest parsing works. (eg. if you set one or more
idempotency attributes via the special splat attribute that references a
variable).